+++ tst/ruCubeIgnoredPiecesGapFiller_test.cpp
@@ -26,7 +26,7 @@ TEST(ruCubeIgnoredPiecesGapFiller, cornersOrientationIgnoredGapsFillNoGapsTest) 
 
     const std::vector<cornersArray> expectedCornersOrient = cornersOrient;
 
-    for (uint8_t i = 0; i < size(cornersOrient); ++i) {
+    for (size_t i = 0; i < size(cornersOrient); ++i) {
         auto orient = cornersOrient[i];
         ASSERT_TRUE(filler.cornersOrientationIgnoredGapsFill(orient));
         ASSERT_EQ(expectedCornersOrient[i], orient);
@@ -66,7 +66,7 @@ TEST(ruCubeIgnoredPiecesGapFiller, cornersOrientationIgnoredGapsFillWithGapsTest
         {  0,  0,  0,  0,  0,  0 },
     };
 
-    for (uint8_t i = 0; i < size(cornersOrient); ++i) {
+    for (size_t i = 0; i < size(cornersOrient); ++i) {
         ASSERT_TRUE(filler.cornersOrientationIgnoredGapsFill(cornersOrient[i]));
         ASSERT_EQ(expectedCornersOrient[i], cornersOrient[i]);
     }
@@ -187,24 +187,13 @@ TEST(ruCubeIgnoredPiecesGapFiller, permutationIgnoredGapsFillNextTest) {
         },
     };
 
-    for (uint8_t i = 0; i < size(permutations); ++i) {
+    for (size_t i = 0; i < size(permutations); ++i) {
         auto& [cp, ep] = permutations[i];
         filler.permutationIgnoredGapsFillInit(cp, ep);
 
         uint16_t j = 0;
         while (filler.permutationIgnoredGapsFillNext(cp, ep)) {
             const auto& [ecp, eep] = expectedPermutations[i][j];
-//            std::cout << "---------------i = " << (int) i << ": " << (int)j << std::endl;
-//            std::cout << "cp and ecp" << std::endl;
-//            std::copy(begin(cp), end(cp), std::ostream_iterator<int>(std::cout, " "));
-//            std::cout << std::endl;
-//            std::copy(begin(ecp), end(ecp), std::ostream_iterator<int>(std::cout, " "));
-//            std::cout << std::endl;
-//            std::cout << "ep and eep" << std::endl;
-//            std::copy(begin(ep), end(ep), std::ostream_iterator<int>(std::cout, " "));
-//            std::cout << std::endl;
-//            std::copy(begin(eep), end(eep), std::ostream_iterator<int>(std::cout, " "));
-//            std::cout << std::endl;
             ASSERT_EQ(ecp, cp);
             ASSERT_EQ(eep, ep);
             ++j;
